+++ src/app/(dashboard)/teacher/achievement/page.tsx
@@ -1,14 +1,44 @@
-import React from 'react'
-import Search from './components/SearchBar'
-import AchievementList from './components/achievementList'
+import styles from "./styles/achievement.module.scss";
+import Search from "./components/SearchBar";
+import AddButton from "./components/buttons/addButton";
+import AchievementList from "./components/achievementList";
+import ImportButton from "./components/buttons/importButton";
+import SortButton from "./components/buttons/sortButton";
 
-export default function page() {
+
+
+export default function AchievementRecordsPage(props: {
+  searchParams?: { query?: string; page?: string };
+}) {
+  const searchParams = props.searchParams || {};
+  const query = searchParams.query || "";
+  const currentPage = Number(searchParams.page) || 1;
+
+  
+
+
   return (
-    <div>
-        <h1>Teacher Achievement Page</h1>
-        <Search placeholder='Search...' />
-        <AchievementList query='' currentPage={1} />
-
+    <div className={styles.pageContainer}>
+      <div className={styles.contentContainer}>
+        <div className={styles.achievementPage}>
+        <h1 style={{ fontWeight: "bold" }}>Achievements Records</h1>
+          <div className={styles.topSection}>
+            <div className={styles.searchAndActions}>
+              <Search placeholder="Search..." />
+              <div className={'${styles.actionButton} ${styles.importButton}'}>
+               <ImportButton />
+           </div>
+           <div className={'${styles.actionButton} ${styles.addButton}'}>
+                <AddButton />
+            </div>
+            <div className={'${styles.actionButton} ${styles.sortButton}'}>
+            <SortButton />
+            </div>
+          </div>
+        </div>
+          <AchievementList query={query} currentPage={currentPage} />
+        </div>
+      </div>
     </div>
-  )
+  );
 }
